Rename versionutil to libversion
