+++ socorro/lib/libversion.py
